roles/kafka/api/requirements.txt
fastapi
uvicorn
confluent-kafka
psycopg2
PyJWT
six

